Fix save from user
