+++ src/administrator/components/com_kunena/Controller/UsersController.php
@@ -119,243 +119,7 @@ class UsersController extends AdminController
 		$this->setRedirect(Route::_("index.php?option=com_kunena&view=user&layout=edit&userid={$userid}", false));
 	}
 
-	/**
-	 * Save
-	 *
-	 * @return  void
-	 *
-	 * @since   Kunena 2.0
-	 *
-	 * @throws  Exception
-	 * @throws  null
-	 */
-	public function save()
-	{
-		if (!Session::checkToken('post'))
-		{
-			$this->app->enqueueMessage(Text::_('COM_KUNENA_ERROR_TOKEN'), 'error');
-			$this->setRedirect(KunenaRoute::_($this->baseurl, false));
 
-			return;
-		}
-
-		$newview      = $this->app->input->getString('newview');
-		$newrank      = $this->app->input->getString('newrank');
-		$signature    = $this->app->input->getString('signature', '');
-		$deleteSig    = $this->app->input->getInt('deleteSig');
-		$moderator    = $this->app->input->getInt('moderator');
-		$uid          = $this->app->input->getInt('uid');
-		$deleteAvatar = $this->app->input->getInt('deleteAvatar');
-		$neworder     = $this->app->input->getInt('neworder');
-		$modCatids    = $moderator ? $this->app->input->get('catid', [], 'array') : [];
-		$modCatids    = ArrayHelper::toInteger($modCatids);
-
-		if ($uid)
-		{
-			$user = KunenaFactory::getUser($uid);
-
-			// Prepare variables
-			if ($deleteSig == 1)
-			{
-				$user->signature = '';
-			}
-			else
-			{
-				$user->signature = $signature;
-			}
-
-			$user->personalText = $this->app->input->getString('personaltext', '');
-			$birthdate          = $this->app->input->getString('birthdate');
-
-			if ($birthdate)
-			{
-				$date = Factory::getDate($birthdate);
-
-				$birthdate = $date->format('Y-m-d');
-			}
-
-			$user->birthdate = $birthdate;
-			$user->location  = trim($this->app->input->getString('location', ''));
-			$user->gender    = $this->app->input->getInt('gender', '');
-			$this->cleanSocial($user, $this->app);
-			$user->websitename  = $this->app->input->getString('websitename', '');
-			$user->websiteurl   = $this->app->input->getString('websiteurl', '');
-			$user->hideEmail    = $this->app->input->getInt('hidemail');
-			$user->showOnline   = $this->app->input->getInt('showonline');
-			$user->canSubscribe = $this->app->input->getInt('cansubscribe');
-			$user->userListtime = $this->app->input->getInt('userlisttime');
-			$user->socialshare  = $this->app->input->getInt('socialshare');
-			$user->view         = $newview;
-			$user->ordering     = $neworder;
-			$user->rank         = $newrank;
-
-			if ($deleteAvatar == 1)
-			{
-				$user->avatar = '';
-			}
-
-			if (!$user->save())
-			{
-				$this->app->enqueueMessage(Text::_('COM_KUNENA_USER_PROFILE_SAVED_FAILED'), 'error');
-			}
-			else
-			{
-				$this->app->enqueueMessage(Text::_('COM_KUNENA_USER_PROFILE_SAVED_SUCCESSFULLY'));
-			}
-
-			$this->setModerate($user, $modCatids);
-		}
-
-		$this->setRedirect(KunenaRoute::_($this->baseurl, false));
-	}
-
-	/**
-	 * Clean social items
-	 *
-	 * @param   KunenaUser  $user user
-	 * @param   Factory     $app  app
-	 *
-	 * @return  void
-	 *
-	 * @since   Kunena 6.0
-	 */
-	protected function cleanSocial(&$user, $app)
-	{
-		foreach ($user->socialButtons() as $key => $social)
-		{
-			$user->$key = str_replace(' ', '', trim($app->input->getString($key, '')));
-		}
-	}
-
-	/**
-	 * Set moderator rights on the user given
-	 *
-	 * @param   KunenaUser  $user      user
-	 * @param   array       $modCatids modCatids
-	 *
-	 * @return  boolean
-	 *
-	 * @since   Kunena 5.1
-	 *
-	 * @throws  Exception
-	 */
-	protected function setModerate(KunenaUser $user, $modCatids)
-	{
-		// Update moderator rights
-		$categories = KunenaForumCategoryHelper::getCategories(false, false, 'admin');
-
-		foreach ($categories as $category)
-		{
-			$category->setModerator($user, in_array($category->id, $modCatids));
-		}
-
-		// Global moderator is a special case
-		if ($this->me->isAdmin())
-		{
-			KunenaAccess::getInstance()->setModerator(0, $user, in_array(0, $modCatids));
-		}
-
-		return true;
-	}
-
-	/**
-	 * Apply
-	 *
-	 * @return  void
-	 *
-	 * @since   Kunena 2.0
-	 *
-	 * @throws  Exception
-	 */
-	public function apply()
-	{
-		if (!Session::checkToken('post'))
-		{
-			$this->app->enqueueMessage(Text::_('COM_KUNENA_ERROR_TOKEN'), 'error');
-
-			return;
-		}
-
-		$newview      = $this->app->input->getString('newview');
-		$newrank      = $this->app->input->getString('newrank');
-		$signature    = $this->app->input->getString('signature', '');
-		$deleteSig    = $this->app->input->getInt('deleteSig');
-		$moderator    = $this->app->input->getInt('moderator');
-		$uid          = $this->app->input->getInt('uid');
-		$deleteAvatar = $this->app->input->getInt('deleteAvatar');
-		$neworder     = $this->app->input->getInt('neworder');
-		$modCatids    = $moderator ? $this->app->input->get('catid', [], 'array') : [];
-		$modCatids    = ArrayHelper::toInteger($modCatids);
-
-		if ($uid)
-		{
-			$user = KunenaFactory::getUser($uid);
-
-			// Prepare variables
-			if ($deleteSig == 1)
-			{
-				$user->signature = '';
-			}
-			else
-			{
-				$user->signature = $signature;
-			}
-
-			$user->personalText = $this->app->input->getString('personaltext', '');
-			$birthdate          = $this->app->input->getString('birthdate');
-
-			if ($birthdate)
-			{
-				$date = Factory::getDate($birthdate);
-
-				$birthdate = $date->format('Y-m-d');
-			}
-
-			$user->birthdate = $birthdate;
-			$user->location  = trim($this->app->input->getString('location', ''));
-			$user->gender    = $this->app->input->getInt('gender', '');
-			$this->cleanSocial($user, $this->app);
-			$user->websitename  = $this->app->input->getString('websitename', '');
-			$user->websiteurl   = $this->app->input->getString('websiteurl', '');
-			$user->hideEmail    = $this->app->input->getInt('hidemail');
-			$user->showOnline   = $this->app->input->getInt('showonline');
-			$user->canSubscribe = $this->app->input->getInt('cansubscribe');
-			$user->userListtime = $this->app->input->getInt('userlisttime');
-			$user->socialshare  = $this->app->input->getInt('socialshare');
-
-			$user->view     = $newview;
-			$user->ordering = $neworder;
-			$user->rank     = $newrank;
-
-			if ($deleteAvatar == 1)
-			{
-				$user->avatar = '';
-			}
-
-			if (!$user->save())
-			{
-				$this->app->enqueueMessage(Text::_('COM_KUNENA_USER_PROFILE_SAVED_FAILED'), 'error');
-			}
-			else
-			{
-				$this->app->enqueueMessage(Text::_('COM_KUNENA_USER_PROFILE_SAVED_SUCCESSFULLY'));
-			}
-
-			// Update moderator rights
-			$categories = KunenaForumCategoryHelper::getCategories(false, false, 'admin');
-
-			foreach ($categories as $category)
-			{
-				$category->setModerator($user, in_array($category->id, $modCatids));
-			}
-
-			// Global moderator is a special case
-			if ($this->me->isAdmin())
-			{
-				KunenaAccess::getInstance()->setModerator(0, $user, in_array(0, $modCatids));
-			}
-		}
-	}
 
 	/**
 	 * Trash menu
